+++ Solution/RiskOfChaos/Content/RoCContent.cs
@@ -334,8 +334,6 @@ namespace RiskOfChaos.Content
             public static GameObject ValueModificationManager;
 
             public static GameObject UIModificationProvider;
-
-            public static GameObject PlayerMoneyToExpConverter;
         }
     }
 }